refcator: 공용 위젯 이름 변경
